Create Theatre Ticket System
This code allows you to see the seats in a theatre and will show you to price. You then can choose to buy a ticket if you do the seat will be removed from the available seats to purchase. If you buy all the tickets it will read sold out.