Theatre Ticket System
This is a code that simulates buying a ticket at a theatre. It shows a top-down view of all the seats in the theatre. It tells you the price of the seat. When all seats are sold out it will say sold out. You can also use a shortcut and type all to buy all seats or type no to leave.